Archived Lumped Actuation Model, started new polar model
The next model is the same physics as the original polar model, but this
time I'm being a bit smarted about how I handle different motion phases
and states. The key difference is that I'm using the minimal
representation of each mode. I'm also only creating three discrete
modes: flight, single stance, and double stance. I realize that this is
a bit more democratic than my previous method. It also minimizes copying
things over and over. Another difference is that is that in single
stance and double stance, one of the feet is always at the origin. I've
also changed my angle definitions.